+++ app_manager_plugin/app_manager_plugin.yaml
