+++ app_manager_plugin/launch/test_stop_plugin.launch
