+++ app_manager_plugin/setup.py
