+++ app_manager_plugin/src/app_manager_plugin/test_start_plugin.py
@@ -1,6 +1,6 @@
 import rospy
 
-from app_manager_plugins import AppManagerPlugin
+from app_manager_plugin import AppManagerPlugin
 
 
 class TestStartPlugin(AppManagerPlugin):
