--- app_manager_plugins/apps/test_app/test_app.app
@@ -1,4 +1,0 @@
-display: Test app
-platform: all
-launch: app_manager_plugins/test_app.xml
-interface: app_manager_plugins/test_app.interface
