--- app_manager_plugins/apps/test_app/test_app.interface
@@ -1,2 +1,0 @@
-published_topics: {}
-subscribed_topics: {}
