--- app_manager_plugins/src/app_manager_plugins/__init__.py
@@ -1,3 +1,0 @@
-from app_manager_plugins.app_manager_plugin import AppManagerPlugin  # NOQA
-from app_manager_plugins.test_start_plugin import TestStartPlugin  # NOQA
-from app_manager_plugins.test_stop_plugin import TestStopPlugin  # NOQA
